Replaced - (unsigned int) count by - (NSUInteger) count in all frameworks to prevent return type conflict with GNUstep Foundation classes.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileFoundation@6128 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/ETClassMirror.m
@@ -261,7 +261,7 @@ which the returned protocols conform to themselves. */
 {
 	return [[self contentArray] objectEnumerator];
 }
-- (unsigned int) count
+- (NSUInteger) count
 {
 	return [[self contentArray] count];
 }
